refactor: 마일리지 리스트 페이지 도메인 구조 변경
