--- src/components/MileageList/index.ts
@@ -1,9 +1,0 @@
-export { default as EmptyMileageTable } from './EmptyMileageTable';
-export { default as JoinedTabs } from './JoinedTabs';
-export { default as MileageCountSection } from './MileageCountSection';
-export { default as MileageFilterSection } from './MileageFilterSection';
-export { default as MileageTable } from './MileageTable';
-export { default as MileageTableListSection } from './MileageTableListSection';
-export { default as MobileMileageCountSection } from './MobileMileageCountSection';
-export { default as SearchMileageInput } from './SearchMileageInput';
-export { default as SemesterDropdown } from './SemesterDropdown';
