+++ src/hooks/index.ts
@@ -1,14 +1,9 @@
 export { default as useEditMileageForm } from './useEditMileageForm';
 export { default as useFile } from './useFile';
 export { default as useFileWithType } from './useFileWithType';
-export { default as useFilteredByJoined } from './useFilteredByJoined';
-export { default as useFilteredByKeyword } from './useFilteredByKeyword';
-export { default as useFilteredBySemester } from './useFilteredBySemester';
-export { default as useGroupedMileageList } from './useGroupedMileageList';
 export { default as useInput } from './useInput';
 export { default as useInputWithValidate } from './useInputWithValidate';
 export { default as useLogin } from './useLogin';
 export { default as useNewMileageForm } from './useNewMileageForm';
 export { default as useOpenModal } from './useOpenModal';
-export { default as useQueryParams } from './useQueryParams';
 export { default as useScholarshipDuration } from './useScholarshipDuration';
