+++ src/hooks/queries/index.ts
@@ -1,7 +1,6 @@
 export { default as useDeleteSubmittedMileageMutation } from './useDeleteSubmittedMileageMutation';
 export { default as useGetEtcMileageQuery } from './useGetEtcMileageQuery';
 export { default as useGetIsAppliedScholarshipQuery } from './useGetIsAppliedScholarshipQuery';
-export { default as useGetMileageQuery } from './useGetMileageQuery';
 export { default as useGetScholarshipDurationQuery } from './useGetScholarshipDurationQuery';
 export { default as useGetSubmittedFileQuery } from './useGetSubmittedFileQuery';
 export { default as useGetSubmittedMileageQuery } from './useGetSubmittedMileageQuery';
