+++ src/pages/MileageListPage/components/JoinedTabs.tsx
@@ -1,9 +1,10 @@
 import { Tabs } from '@/components';
-import { filterJoinedItems } from '@/constants/filterItems';
-import { useFilteredByJoined } from '@/hooks';
 import { TabItem } from '@/types/tab';
 
-const JoinedTabs = () => {
+import { filterJoinedItems } from '../constants/filterItems';
+import { useFilteredByJoined } from '../hooks/useFilteredByJoined';
+
+export const JoinedTabs = () => {
   const { selectedJoined, setSelectedJoined } = useFilteredByJoined();
 
   const handleChange = (newItem: TabItem) => {
@@ -18,5 +19,3 @@ const JoinedTabs = () => {
     />
   );
 };
-
-export default JoinedTabs;
