+++ src/pages/MileageListPage/constants/system.ts
@@ -1,0 +1,1 @@
+export const ALL_SEMESTER = '모든 학기';
