+++ src/pages/MileageListPage/hooks/useFilteredByKeyword.ts
@@ -1,8 +1,9 @@
-import { useQueryParams } from '@/hooks';
 import { debounce } from '@/utils/debounce';
 import { useCallback } from 'react';
 
-const useFilteredByKeyword = () => {
+import { useQueryParams } from './useQueryParams';
+
+export const useFilteredByKeyword = () => {
   const { queryParams, updateQueryParams } = useQueryParams();
   const keyword = queryParams.keyword;
 
@@ -15,5 +16,3 @@ const useFilteredByKeyword = () => {
 
   return { keyword, setKeyword };
 };
-
-export default useFilteredByKeyword;
